weeks/week-05/app/requirements.txt
fastapi==0.104.1
strawberry-graphql==0.221.0
uvicorn==0.24.0